End run, reset & School World split option added
If the player chooses to let the game play the ending cutscene after collecting Tower World's dream shard, it'll automatically split (to end the run) and deduct 6 seconds from their time since it takes exactly that long for the cutscene to start loading.

The game now resets the timer if the player exits to the main menu from the Dream Nexus (but not the Dream Hub) or from Heaven. It doesn't reset on exiting the game, since that's required in the All Achievements run.

There's also an option to split upon finishing School World - I'm not sure why I didn't have that in before? Or why splitting on entering School World was default...